% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{build_manuals}
\alias{build_manuals}
\title{Build manuals (including vignettes) in the working copy.}
\usage{
build_manuals(pkg = getwd())
}
\arguments{
\item{pkg}{The directory holding the working copy.}
}
\description{
Build manuals (including vignettes) in the working copy.
}
