% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_testnames}
\alias{check_testnames}
\title{Check test names against pattern.}
\usage{
check_testnames(pkg = getwd())
}
\arguments{
\item{pkg}{The directory holding the working copy.}
}
\description{
Check test names against pattern.
}
