% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{clean_and_convert_layout}
\alias{clean_and_convert_layout}
\title{Convert layout from old to new format. Can be safely applied to new layouts.}
\usage{
clean_and_convert_layout(df, is_demo = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{is_demo}{Whether to apply further normalization to minimize entries.}
}
\description{
Convert layout from old to new format. Can be safely applied to new layouts.
}
