% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSS.R
\name{compute_dss}
\alias{compute_dss}
\title{Compute a Drug Sensitivity Score (DSS)}
\usage{
compute_dss(
  pars_table,
  resp_min = 10,
  score_type = 3,
  conc_scale_unused = 1e-09,
  apply_log_to_conc = TRUE,
  curve = NA,
  force_symm = FALSE,
  resp_max = 100,
  bottom_corr = TRUE,
  allow_big_mid = FALSE,
  int_decreasing = FALSE,
  allow_nan = TRUE
)
}
\arguments{
\item{pars_table}{can be a data.frame of (in that order!)}

\item{resp_min}{TODO}

\item{score_type}{TODO}

\item{conc_scale_unused}{TODO}

\item{apply_log_to_conc}{TODO}

\item{curve}{TODO}

\item{force_symm}{TODO}

\item{resp_max}{TODO}

\item{bottom_corr}{TODO}

\item{allow_big_mid}{TODO}

\item{int_decreasing}{TODO}

\item{allow_nan}{TODO}
}
\description{
Compute a Drug Sensitivity Score (DSS)
}
