% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{debug_save}
\alias{debug_save}
\title{Save arguments in rds files named after arguments.}
\usage{
debug_save(...)
}
\arguments{
\item{...}{Variables to be saved (not quoted).}
}
\description{
Save arguments in rds files named after arguments.
}
