% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{prep_and_check}
\alias{prep_and_check}
\title{Prepare and check iTReX in the working copy.}
\usage{
prep_and_check(pkg = getwd(), strict = FALSE)
}
\arguments{
\item{pkg}{The directory holding the working copy.}

\item{strict}{Whether to check for CRAN and BioConductor compatibility.}
}
\description{
Prepare and check iTReX in the working copy.
}
