% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{prepare}
\alias{prepare}
\title{Prepare working copy for commit.}
\usage{
prepare(pkg = getwd(), manual = TRUE)
}
\arguments{
\item{pkg}{The directory holding the working copy.}

\item{manual}{Whether to run \code{\link[=build_manuals]{build_manuals()}}.}
}
\description{
Prepare working copy for commit.
}
