% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{prepare_demos}
\alias{prepare_demos}
\title{Prepare demo data (xlsx and zip) in the working copy.}
\usage{
prepare_demos(pkg = getwd())
}
\arguments{
\item{pkg}{The directory holding the working copy.}
}
\description{
Prepare demo data (xlsx and zip) in the working copy.
}
