% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{record_test}
\alias{record_test}
\title{Record a new test and store it in the working copy.}
\usage{
record_test(pkg = getwd())
}
\arguments{
\item{pkg}{The directory holding the working copy.}
}
\description{
Record a new test and store it in the working copy.
}
