% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{show_system_deps}
\alias{show_system_deps}
\title{Generate system dependencies for README.md}
\usage{
show_system_deps(pkg = getwd())
}
\arguments{
\item{pkg}{The directory holding the working copy.}
}
\description{
Generate system dependencies for README.md
}
