% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{test}
\alias{test}
\title{Run iTReX tests}
\usage{
test(testnames = NULL, quiet = FALSE, url = NULL)
}
\arguments{
\item{testnames}{Passed to \code{\link[shinytest:testApp]{shinytest::testApp()}}}

\item{quiet}{Passed to \code{\link[shinytest:testApp]{shinytest::testApp()}}}

\item{url}{URL of test-mode Shiny instance to connect to.
Requires a customized {shinytest} instance as of {shinytest} v1.5.1.}
}
\description{
Run iTReX tests
}
\examples{
test(testnames = c("tabs_home"))
\dontshow{
# https://github.com/r-lib/processx/issues/294
processx::supervisor_kill()
}
}
