% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{test_and_update}
\alias{test_and_update}
\title{Run (and update) tests in the working copy.}
\usage{
test_and_update(..., pkg = getwd())
}
\arguments{
\item{...}{Parameters (such as testnames) passed to [shinytest::testApp()].}

\item{pkg}{The directory holding the working copy.}
}
\description{
Run (and update) tests in the working copy.
}
