% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{update_tests}
\alias{update_tests}
\title{Update test-expected in the working copy.}
\usage{
update_tests(..., pkg = getwd())
}
\arguments{
\item{...}{Parameters passed to [shinytest::snapshotUpdate()].}

\item{pkg}{The directory holding the working copy.}
}
\description{
Update test-expected in the working copy.
}
