% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{version}
\alias{version}
\title{Report iTReX version}
\usage{
version(long = FALSE)
}
\arguments{
\item{long}{Whether to print the full version string with git commit hash.}
}
\description{
Report iTReX version
}
\examples{
version()
version(long = TRUE)
}
