% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{test}
\alias{test}
\title{Run iTReX tests}
\usage{
test(testnames = NULL, quiet = FALSE)
}
\arguments{
\item{testnames}{Passed to [shinytest::testApp]}

\item{quiet}{Passed to [shinytest::testApp]}
}
\description{
Run iTReX tests
}
\examples{
test(testnames = c("tabs_home"))
\dontshow{
# https://github.com/r-lib/processx/issues/294
processx::supervisor_kill()
}
}
